+++ src/datahandling/expressionparser.h
@@ -12,16 +12,19 @@ public:
 
     QList<DataPoint> dataPoints() const;
     QStringList processedExpressions() const;
+    QList<QList<int>> expressionDataPointIndices() const;
 
 private:
     void parseExpressions(const QStringList& expressions);
 
-    QString processExpression(QString const& expr);
+    QString processExpression(QString const& expr, QList<int>& dataPointIndices);
     bool processRegisterExpression(const QString& regExpr, DataPoint& dataPoint);
-    QString constructInternalRegisterFunction(DataPoint const& dataPoint, int size);
+    int findOrAddDataPoint(DataPoint const& dataPoint);
+    QString constructInternalRegisterFunction(int idx, int size);
 
     QStringList _processedExpressions;
     QList<DataPoint> _dataPoints;
+    QList<QList<int>> _expressionDataPointIndices;
 
     QRegularExpression _findRegRegex;